django_backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
web3>=6.0
pyserial>=3.5
